Add support for syncing from HTTP
Sources that use HTTP or HTTPS should now be syncable using this
tool to Amazon S3. Tested using an S3 HTTPS URL as a source, first
download/upload works as expected, second run will exit cleanly as
the source does not have a file newer than the destination (using
the 'If-Modified-Since' HTTP header).